CLI: Replace path.join with path.resolve
This fixes an issue where absolute paths couldn't be used
with the -t option.

path.resolve is a lot cleverer than path.join (which is
rather naive). For example:

  path.join('foo', '/bar')    -> 'foo/bar'
  path.resolve('foo', '/bar') -> '/bar'

path.resolve also defaults to using the current working
directory if no 'from' argument is supplied, so we can
leave that out.